refactor: client into modules closes #85
