+++ .changeset/nine-needles-know.md
@@ -1,0 +1,5 @@
+---
+"@ledgerhq/wallet-api-client": minor
+---
+
+refactor: client into modules
